requirements.txt
duckdb
sentence-transformers
tqdm
python-dotenv
